pyproject.toml
[tool.poetry]
name = "mcsqs"
version = "0.1.0"
description = "Enables multiple SQS queues to read from a single SQS queue using SNS."
authors = ["awhipp <alexruns4@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
boto3 = "^1.34.83"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
